Move validators to correct folder
